backend/fast/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-dotenv
python-jose
sqlalchemy[asyncio]
pydantic
passlib[bcrypt]
asyncpg
google-generativeai